fix(#1119): Consolidate circuit-breaker implementations into shared package
- Create unified CircuitBreaker in packages/stellar/src/circuit-breaker.ts
- Support both functional wrapper API (call()) and manual state API (isOpen(), recordSuccess(), recordFailure())
- Update backend circuit-breaker to re-export from shared implementation
- Update horizon-client to use shared CircuitBreaker with proper configuration mapping
- Ensures all circuit breaker instances share same behavior and bug fixes
